feat: Add current time endpoint to AppController and fix typo in NewDataSourceForm
This commit adds a new endpoint to the AppController that returns the current time. It also fixes a typo in the NewDataSourceForm component where "can't" was misspelled as "can&apos;t".

Changes:
- Added the Any() method to the AppController that returns the current time.
- Fixed the typo in the FormDescription of the NewDataSourceForm component.

This feature improves the functionality of the AppController by providing an endpoint to retrieve the current time and enhances the user experience by correcting a typo in the NewDataSourceForm component.