refactor: Add email field to GithubStrategy
This commit adds an email field to the GithubStrategy in the auth module. The email is extracted from the user's profile and used to update the user's email in the database. If the email is not available in the profile, a default email address is used.

Changes:
- Added email field to the GithubStrategy in auth module.
- Extracted email from user's profile and updated user's email in the database.
- If email is not available in the profile, a default email address is used.

This refactor improves the functionality of the Github authentication strategy by including the user's email in the authentication process.