feat: Add toggleSync endpoint to SourceController
This commit adds a new endpoint `toggleSync` to the `SourceController` in the `api_gateway/src/source/source.controller.ts` file. The `toggleSync` endpoint allows users to toggle the synchronization status of a data source connection. It takes the `id` of the connection as a parameter and updates the `syncOn` field of the corresponding `DataSourceConnections` model in the database.

Changes:
- Added `toggleSync` endpoint to `SourceController` in `api_gateway/src/source/source.controller.ts`.
- The `toggleSync` endpoint allows users to toggle the synchronization status of a data source connection.

This feature enhances the functionality of the API Gateway service by providing a convenient way for users to control the synchronization of their data source connections. latest